+++ src/components/MTls/MTLSStatus.tsx
@@ -9,13 +9,6 @@ type Props = {
   overlayPosition?: string;
 };
 
-export enum MTLSStatuses {
-  ENABLED = 'MTLS_ENABLED',
-  PARTIALLY = 'MTLS_PARTIALLY_ENABLED',
-  NOT_ENABLED = 'MTLS_NOT_ENABLED',
-  DISABLED = 'MTLS_DISABLED'
-}
-
 export type StatusDescriptor = {
   message: string;
   icon: string;
